Alpha release - Non-validator PulseChain node setup

+++ helper/tmux_logviewer.sh
@@ -1,34 +1,28 @@
 #!/bin/bash
 
-# Kill prev. sessions
-tmux kill-session -t logs
+# Extract logs in a tmux session
 
-# Create a new tmux session named 'logs'
-tmux new-session -d -s logs
+# Check if a tmux session named "logs" already exists
+if tmux has-session -t logs 2>/dev/null; then
+    # Kill the existing session
+    tmux kill-session -t logs
+fi
 
-# Split the window horizontally
+# Create a new tmux session named "logs" with a window named "execution"
+tmux new-session -d -s logs -n "execution"
+
+# Split the window horizontally and create a pane for the beacon logs
 tmux split-window -h -t logs
 
-# Select the first pane and split it vertically
-tmux split-window -v -t logs:0.0
+# Name the left pane "execution" and the right pane "beacon"
+tmux send-keys -t logs:0.0 'echo -e "\033]0;Execution Client Logs\007"' Enter
+tmux send-keys -t logs:0.1 'echo -e "\033]0;Beacon Client Logs\007"' Enter
 
-# Select the second pane and split it vertically
-tmux split-window -v -t logs:0.1
-
-# Evenly distribute the pane sizes
-tmux select-layout -t logs tiled
-
-# Send the commands to the respective panes
+# Start tailing the logs in each pane
 tmux send-keys -t logs:0.0 'docker logs -f --tail=30 execution' Enter
-tmux send-keys -t logs:0.1 'docker logs -f --tail=30 validator' Enter
-tmux send-keys -t logs:0.2 'docker logs -f --tail=30 beacon' Enter
-tmux send-keys -t logs:0.3 'htop' Enter
-
-# Create a new window with 'sudo watch df -H' running
-tmux new-window -n "DiskUsage" -t logs:1 'watch df -H'
-
-# Select the first pane (0.0)
-tmux select-pane -t logs:0.0
+tmux send-keys -t logs:0.1 'docker logs -f --tail=30 beacon' Enter
 
 # Attach to the tmux session
-tmux attach-session -t logs:0.0
+tmux attach-session -t logs
+
+# Note: To exit the session, press Ctrl+B followed by :, then type "kill-session" and press Enter
